fix: extract getGenericExplorerUrl to explorer helper
It can't be in main helpers/utils as it causes circular dependency.
